Created class ExportUsfm for ease of testing.
